// app/types.ts
export interface LoginType {
  mode: "login" | "signup" | "password" | "admin" | "advert";
}